+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * feat(`MainTransformer`): Manual garbage collector.
 * feat(`ObjectToObjectTransformer`): Option to disable lazy loading.
 * feat(`ObjectToObjectTransformer`): Option to disable target value reading.
+* refactor(`MainTransformer`): Make manual GC interval a static variable.
 
 ## 0.7.3